server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
mysql-connector-python==9.3.0
packaging==25.0
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
yarl==1.20.0
certifi==2025.4.26
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-asyncio==0.26.0
